+++ ios/Classes/lib/notifications/models/NotificationIntervalModel.swift
@@ -57,6 +57,10 @@ public class NotificationIntervalModel : NotificationScheduleModel {
         if(IntUtils.isNullOrEmpty(interval) || interval! <= 0){
             throw AwesomeNotificationsException.invalidRequiredFields(msg: "Interval cannot be null, empty or zero")
         }
+
+        if((repeats ?? false) && interval! < 60){
+            throw AwesomeNotificationException.invalidRequiredFields(msg: "time interval must be at least 60 if repeating");
+        }
     }
     
     public func getUNNotificationTrigger() -> UNNotificationTrigger? {