+++ lib/src/models/notification_interval.dart
@@ -58,5 +58,11 @@ class NotificationInterval extends NotificationSchedule {
   @override
   void validate() {
     assert((this.interval ?? -1) >= 0);
+
+    assert(
+        this.repeats &&
+        (this.interval ?? 0) >= 60,
+        'time interval must be at least 60 if repeating'
+    );
   }
 }